Move action selector block from main.rb
