+++ main.rb
@@ -1,4 +1,5 @@
 require './app'
+require './action-selector'
 
 def display_menu
   puts 'Please choose an option by entering a number:'
@@ -12,30 +13,15 @@ def display_menu
   print 'Your input: '
 end
 
-def action_selector(input, app)
-  case input
-  when 1
-    app.list_all_books
-  when 2
-    app.list_people
-  when 3
-    app.create_person
-  when 4
-    app.create_book
-  when 5
-    app.create_rental
-  when 6
-    app.list_rental_per_person
-  end
-end
 
 def main
   app = App.new
+  action = ActionSeclector.new
   loop do
     display_menu
     input_number = gets.chomp.to_i
     if input_number.positive? and input_number < 7
-      action_selector(input_number, app)
+      action.action_selector(input_number, app)
     elsif input_number == 7
       puts 'Goodbye see you soon!'
     else
